Create technical writing article comparing Java and JavaScript
First of a series of technical articles for my developer portfolio. 

This article is part of my portfolio to practice technical writing. It compares Java and JavaScript in terms of syntax, execution, and usage to clarify common misconceptions. It will later be used in my blog and GitHub Pages site.